Modify: functions.py file split into three
The functions for creating data-structures were moved into create.py

The functions for welcome-message, and intro were moved into welcome.py

Reason: to make code more readable and not clutter up the main functions.py. Now, when the code for say creating a new verb needs to be changed it takes place in the create.py file.